Score is not an int

+++ addok/core.py
@@ -190,7 +190,7 @@ class Result(object):
     @property
     def score(self):
         if self._score != '':
-            return int(self._score)
+            return float(self._score)
         score, _max = zip(*self._scores.values())
         return sum(score) / sum(_max)
